requirements.txt
tweepy
google-generativeai
matplotlib
yfinance
feedparser
python-dotenv
pandas